Consolidate serviceLinks and osfServices into the same file.
