+++ addon/components/navbar-auth-dropdown/component.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import layout from './template';
 import config from 'ember-get-config';
-import serviceLinks from '../../const/service-links';
+import { serviceLinks } from '../../const/service-links';
 
 /**
  * @module ember-osf
