+++ addon/const/service-links.js
@@ -34,4 +34,38 @@ const serviceLinks = {
     settings: `${osfUrl}settings/`
 };
 
-export default serviceLinks;
+
+/**
+ * @class osf-services
+ */
+
+/**
+ * Provides list of OSF services and their links
+ * @property osfServices
+ * @final
+ * @type {Array}
+ */
+
+const osfServices = [
+    {
+        name: 'HOME',
+        url: serviceLinks.osfHome,
+    },
+    {
+        name: 'PREPRINTS',
+        url: serviceLinks.preprintsHome
+    },
+    {
+        name: 'REGISTRIES',
+        url: serviceLinks.registriesHome
+    },
+    {
+        name: 'MEETINGS',
+        url: serviceLinks.meetingsHome
+    }
+];
+
+export {
+    serviceLinks,
+    osfServices
+};
